move funcs to utils, add some comments
